% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/file.R
\name{get_file}
\alias{get_file}
\title{Get a Google Doubleclick API report file}
\usage{
get_file(path, token)
}
\arguments{
\item{path}{list including a request parameters.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}

\item{query}{list including a request parameters.}
}
\value{
A Google Doubleclick API file response.
}
\description{
Get a Google Doubleclick API report file
}

