% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/request.R
\name{get_response}
\alias{get_response}
\title{Get a Google Doubleclick API response}
\usage{
get_response(type = c("ga", "rt", "mcf", "mgmt", "dfa"), path = NULL,
  query = NULL, reportId = NULL, reportCMD = NULL, fileId = NULL,
  simplify = TRUE, flatten = TRUE, token)
}
\arguments{
\item{type}{character string including report type.}

\item{path}{list including a request parameters.}

\item{query}{list including a request parameters.}

\item{simplify}{logical. Coerce JSON arrays to a vector, matrix or data frame.}

\item{flatten}{logical. Automatically flatten nested data frames into a single non-nested data frame.}

\item{token}{\code{\link[httr]{Token2.0}} class object with a valid authorization data.}
}
\value{
A list contatin Google Doubleclick API response.
}
\description{
Get a Google Doubleclick API response
}

